requirements.txt
aiohttp==2.2.5
appnope==0.1.0
async-timeout==1.4.0
attrdict==2.0.0
attrs==17.2.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coverage==4.4.1
cytoolz==0.8.2
decorator==4.1.2
ecdsa==0.13
ethereum-abi-utils==0.4.3
ethereum-utils==0.4.1
hypothesis==3.21.0
hypothesis-regex==0.2
idna==2.6
inflection==0.3.1
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.11.0
jsonref==0.1
jsonschema==2.6.0
Markdown==2.4
multidict==3.2.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pylru==1.0.9
pysha3==1.0.2
pytest==3.2.2
python-jsonschema-objects==0.2.2
PyYAML==3.12
requests==2.18.4
rlp==0.6.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.8.2
trafaret==0.12.0
git+https://github.com/8gen/trafaret_schema
trafaret-schema==0.1.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
web3==3.15.0
yarl==0.12.0
